+++ account-service/src/main/java/com/example/accountservice/service/AccHolderService.java
@@ -4,11 +4,11 @@ import com.example.accountservice.model.AccHolder;
 
 public interface AccHolderService {
 
-    AccHolder findById(int id);
+    AccHolder findById(Long id);
 
     AccHolder save(AccHolder accHolder);
 
     AccHolder update(AccHolder accHolder);
 
-    void deleteById(int id);
+    void deleteById(Long id);
 }
